Add an option to produce a separate CSS file (fix #6) and minification (fix #7)
By default, CSS is inlined into the JS bundle and injected on execution.
This commit adds an option (`separateCSS`) that allows to produce a separate CSS
file, as available in the main CSS plugin[1].
The code is heavily influenced by the main CSS plugin, so it also uses
clean-css[2] under the hood to produce the CSS file.

One nice side effect of using clean-css is that we can also optionally minify
the CSS, using the Jspm CLI option --minify/-m
> jspm bundle ... --minify

CSS source maps are also supported thanks to clean-css.

[1] https://github.com/systemjs/plugin-css
[2] https://github.com/jakubpawlowicz/clean-css